Added Timer A
Added timers.c and timers.h to control Timer A0. Timer A0 is configured to set CCR0 to an increment value of one blink per minute (32768/60) which was calculated based on CLK A (32768 Hz).  By setting Timer A in Up mode, the TA0_0_IRQHandler is called every time the TAR value hits CCR0. A count variable is incremented every time the ISR is called. The LED is turned on slightly before the blink_rate count is reached. The count value is reset and LED turns off when it reaches blink_rate blinks per minute.