USB Dongle Communication with MSP over UART :confetti_ball:
The USB dongle now communicates with the MSP over UART A0. The dongle takes user input and then sends the desired bpm to the MSP (number + \0). The MSP then looks for the \0 end character and changes the blink rate based on the number before it. The MSP then echos back the bpm (number + \0). The codeblocks program prints all communication.